refactor: Exception 구조 수정
