+++ docs/TODO.md
@@ -2,7 +2,7 @@
 - [x] 초기값이 null인 변수를 생성자 체이닝으로 null을 초기화해도 되는지 알아보기
 - [x] Line 관련 실패하는 E2E 테스트 만들기
 - [x] Line 관련 테스트 코드 리팩토링
-- [ ] Exception 계층 구조 만들기
+- [x] Exception 계층 구조 만들기
 - [ ] Station, Line 없는 입력값 검증하기
 - [x] CustomException 만들기
 - [x] 누락된 테스트 작성하기
