--- src/main/java/wooteco/subway/GlobalExceptionHandler.java
@@ -1,21 +1,0 @@
-package wooteco.subway;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import wooteco.subway.exception.ClientRuntimeException;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(ClientRuntimeException.class)
-    public ResponseEntity<String> handleClientException(final ClientRuntimeException e) {
-        return ResponseEntity.status(e.getHttpStatus()).body(e.getMessage());
-    }
-
-    @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity<String> handleServerException(final RuntimeException e) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-    }
-}
