+++ src/main/java/wooteco/subway/exception/DataNotFoundException.java
@@ -2,9 +2,9 @@ package wooteco.subway.exception;
 
 import org.springframework.http.HttpStatus;
 
-public class DataNotFoundException extends ClientRuntimeException {
+public class DataNotFoundException extends RuntimeException {
 
     public DataNotFoundException(final String message) {
-        super(message, HttpStatus.NOT_FOUND);
+        super(message);
     }
 }
