+++ src/main/java/wooteco/subway/exception/DuplicatedNameException.java
@@ -2,9 +2,9 @@ package wooteco.subway.exception;
 
 import org.springframework.http.HttpStatus;
 
-public class DuplicatedNameException extends ClientRuntimeException {
+public class DuplicatedNameException extends RuntimeException {
 
     public DuplicatedNameException(final String message) {
-        super(message, HttpStatus.BAD_REQUEST);
+        super(message);
     }
 }
