--- src/main/java/wooteco/subway/line/DuplicatedStationException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.line;
-
-public class DuplicatedStationException extends RuntimeException {
-
-    public DuplicatedStationException(final String message) {
-        super(message);
-    }
-}
