--- src/main/java/wooteco/subway/line/SectionSizeException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.line;
-
-public class SectionSizeException extends RuntimeException{
-
-    public SectionSizeException(final String message) {
-        super(message);
-    }
-}
