+++ src/main/java/wooteco/subway/line/exception/BiggerDistanceException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.line.exception;
+
+public class BiggerDistanceException extends ClientRuntimeException {
+
+    public BiggerDistanceException(final String message) {
+        super(message);
+    }
+}
