+++ src/main/java/wooteco/subway/line/exception/BothExistentStationException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.line.exception;
+
+public class BothExistentStationException extends ClientRuntimeException {
+
+    public BothExistentStationException(final String message) {
+        super(message);
+    }
+}
