+++ src/main/java/wooteco/subway/line/exception/ClientRuntimeException.java
@@ -1,0 +1,10 @@
+package wooteco.subway.line.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class ClientRuntimeException extends RuntimeException {
+
+    public ClientRuntimeException(final String message) {
+        super(message);
+    }
+}
