+++ src/main/java/wooteco/subway/line/exception/DuplicatedStationException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.line.exception;
+
+public class DuplicatedStationException extends ClientRuntimeException {
+
+    public DuplicatedStationException(final String message) {
+        super(message);
+    }
+}
