+++ src/main/java/wooteco/subway/line/exception/NoneExistentStationException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.line.exception;
+
+public class NoneExistentStationException extends ClientRuntimeException {
+
+    public NoneExistentStationException(final String message) {
+        super(message);
+    }
+}
