+++ src/main/java/wooteco/subway/line/exception/SectionSizeException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.line.exception;
+
+public class SectionSizeException extends ClientRuntimeException {
+
+    public SectionSizeException(final String message) {
+        super(message);
+    }
+}
