--- src/main/java/wooteco/subway/line/section/BiggerDistanceException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.line.section;
-
-public class BiggerDistanceException extends RuntimeException {
-
-    public BiggerDistanceException(final String message) {
-        super(message);
-    }
-}
