--- src/main/java/wooteco/subway/line/section/BothExistentStationException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.line.section;
-
-public class BothExistentStationException extends RuntimeException {
-
-    public BothExistentStationException(final String message) {
-        super(message);
-    }
-}
