--- src/main/java/wooteco/subway/line/section/NoneExistentStationException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.line.section;
-
-public class NoneExistentStationException extends RuntimeException {
-
-    public NoneExistentStationException(final String message) {
-        super(message);
-    }
-}
