+++ src/test/java/wooteco/subway/line/section/SectionTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.line.exception.BiggerDistanceException;
 
 class SectionTest {
