+++ src/test/java/wooteco/subway/line/section/SectionsTest.java
@@ -9,6 +9,8 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.line.exception.BothExistentStationException;
+import wooteco.subway.line.exception.NoneExistentStationException;
 
 public class SectionsTest {
